LayoutTests/imported/w3c/web-platform-tests/custom-elements/form-associated/ElementInternals-accessibility-expected.txt


PASS role is defined in ElementInternals
FAIL ariaActiveDescendantElement is defined in ElementInternals assert_inherits: property "ariaActiveDescendantElement" not found in prototype chain
PASS ariaAtomic is defined in ElementInternals
PASS ariaAutoComplete is defined in ElementInternals
PASS ariaBusy is defined in ElementInternals
PASS ariaChecked is defined in ElementInternals
PASS ariaColCount is defined in ElementInternals
PASS ariaColIndex is defined in ElementInternals
PASS ariaColSpan is defined in ElementInternals
FAIL ariaControlsElements is defined in ElementInternals assert_inherits: property "ariaControlsElements" not found in prototype chain
PASS ariaCurrent is defined in ElementInternals
FAIL ariaDescribedByElements is defined in ElementInternals assert_inherits: property "ariaDescribedByElements" not found in prototype chain
FAIL ariaDetailsElements is defined in ElementInternals assert_inherits: property "ariaDetailsElements" not found in prototype chain
PASS ariaDisabled is defined in ElementInternals
FAIL ariaErrorMessageElement is defined in ElementInternals assert_inherits: property "ariaErrorMessageElement" not found in prototype chain
PASS ariaExpanded is defined in ElementInternals
FAIL ariaFlowToElements is defined in ElementInternals assert_inherits: property "ariaFlowToElements" not found in prototype chain
PASS ariaHasPopup is defined in ElementInternals
PASS ariaHidden is defined in ElementInternals
PASS ariaInvalid is defined in ElementInternals
PASS ariaKeyShortcuts is defined in ElementInternals
PASS ariaLabel is defined in ElementInternals
FAIL ariaLabelledByElements is defined in ElementInternals assert_inherits: property "ariaLabelledByElements" not found in prototype chain
PASS ariaLevel is defined in ElementInternals
PASS ariaLive is defined in ElementInternals
PASS ariaModal is defined in ElementInternals
PASS ariaMultiLine is defined in ElementInternals
PASS ariaMultiSelectable is defined in ElementInternals
PASS ariaOrientation is defined in ElementInternals
FAIL ariaOwnsElements is defined in ElementInternals assert_inherits: property "ariaOwnsElements" not found in prototype chain
PASS ariaPlaceholder is defined in ElementInternals
PASS ariaPosInSet is defined in ElementInternals
PASS ariaPressed is defined in ElementInternals
PASS ariaReadOnly is defined in ElementInternals
PASS ariaRelevant is defined in ElementInternals
PASS ariaRequired is defined in ElementInternals
PASS ariaRoleDescription is defined in ElementInternals
PASS ariaRowCount is defined in ElementInternals
PASS ariaRowIndex is defined in ElementInternals
PASS ariaRowSpan is defined in ElementInternals
PASS ariaSelected is defined in ElementInternals
PASS ariaSort is defined in ElementInternals
PASS ariaValueMax is defined in ElementInternals
PASS ariaValueMin is defined in ElementInternals
PASS ariaValueNow is defined in ElementInternals
PASS ariaValueText is defined in ElementInternals
